Don't clean Prefetch
long: Prefetch is a component in Windows which stores files for quicker launches of programs. basically 'boot cache' to be simplified.  clearing this out requires Windows to do it all again. the only reason you would ever need to refresh the Prefetch files is if the program got a massive overhaul update including startup. However I'm not including it in what is supposed to be a weekly or monthly cleaning