Windows Defender!
long: Windows Defender is the built in antivirus for Windows 10 and 11. Very helpful. However it is somewhat a resource hog, can suddenly spike in resources used, and is a huge privacy concern. Really anything owned by Microsoft is. However it looks through and scans all your files which makes it worse. Not recommended to remove, however if you're focused on privacy and/or performance and you know what you're doing, that option is there.